Updates for rename of SDIMetadata to SocatMetadata
