+++ .idea/modules.xml
@@ -4,7 +4,7 @@
     <modules>
       <module fileurl="file://$PROJECT_DIR$/DashboardApps/DashboardApps.iml" filepath="$PROJECT_DIR$/DashboardApps/DashboardApps.iml" />
       <module fileurl="file://$PROJECT_DIR$/OmeMetadata/OmeMetadata.iml" filepath="$PROJECT_DIR$/OmeMetadata/OmeMetadata.iml" />
-      <module fileurl="file://$PROJECT_DIR$/../SDIMetadata/SDIMetadata.iml" filepath="$PROJECT_DIR$/../SDIMetadata/SDIMetadata.iml" />
+      <module fileurl="file://$PROJECT_DIR$/../SocatMetadata/SocatMetadata.iml" filepath="$PROJECT_DIR$/../SocatMetadata/SocatMetadata.iml" />
       <module fileurl="file://$PROJECT_DIR$/UploadDashboard/UploadDashboard.iml" filepath="$PROJECT_DIR$/UploadDashboard/UploadDashboard.iml" />
     </modules>
   </component>
